[CIR] LLVM lowering support for pointers to member functions
This patch adds support for LLVM lowering of pointers to member functions. The
lowering is ABI-specific and this patch only considers Itanium ABI.

Itanium ABI lowers pointers to member functions to a struct with two fields of
type `ptrdiff_t`. To extract fields from such aggregate values, this patch
includes a new operation `cir.extract_member` to accomplish this.