[CIR] Add attribute visitor for lowering globals
This adds a new mlir-tablegen option to generate a .inc file with
the complete set of attrdefs defined in a .td file and uses the
file generated for CIR attrdefs to create an attr visitor. This visitor
is used in the lowering of global variables directly to LLVM IR.

The purpose of this change is to align the incubator lowering
implementation with the recent upstream changes to make future
upstreaming easier, while also fulfilling the upstream request to
have the visitor be based on a tablegen created file.

The new mlir-tablegen feature will be upstreamed after it is established
here.