[CIR] Delete `FuncOp::verifyType`
Get rid of the function `FuncOp::verifyType`.  The function performed
three checks:

1. Check that `isa<cir::FuncType>(getFunctionType())`.  This is a
tautology that is always true, since the return type of
`getFunctionType()` is already `cir::FuncType`.

2. Report an error if `type.isVarArg() && type.getNumInputs() == 0`,
i.e. when a variadic function has no named parameters.  That check is
incorrect.  In C++, variadic functions don't need to have any named
parameters.  `void f(...) { }` is legal in C++ and ClangIR needs to be
able to compile it.

3. Report an error when the return type is `void`.  This check is
correct (`void` return is represented as no return in MLIR), but it is
redundant.  This is already checked in `FuncType::verify`.

Since `FuncOp::verifyType` serves no useful purpose, delete it, along
with the test for `int variadic(...)` that was in
`clang/test/CIR/IR/invalid.cir`.