[CIR][Lowering][MLIR] Generalize the lowering of cir.ptr_stride
Handle pointer strides in quite more contexts.
Make the code type-safe with the right memref layouts matching the
reinterpret_cast details to avoid tripping the verifiers.
This problem was hidden in the past due to the fact that the load/store lowering
discarded some intermediate incorrect code for the tested simple cases.