[CIR][MLIR] Fix overloaded virtual warnings
There is a difference in behavior between clang and gcc that was leading to
a very large number of -Woverloaded-virtual warnings when compiling with
gcc.

The difference in behavior is that gcc warns for "partial" virtual overloads.
That is, when a base class contains multiple virtual functions with the same
name and a derived class overrides at least one but not all of them, gcc
issues a warning saying that the non-overridden variations were hidden. This
diagnostic is only reported if the -Woverloaded-virtual is explicitly
specified on the command line, which it is by the clang project.

The standard way to suppress this warning, indicating that the behavior was
intentional (which it was in all cases I'm fixing here) is to add a
"using" statement for the overloaded symbol in the derived class. However,
there was a secondary problem that this couldn't be done in the ClangIR
classes because of a "using" clause in the base class that was marked as
"private".

This change updates a few files in the MLIR, both to allow subclasses to add
their own "using" statements by making the base class "using" statement
protected, and also fixes a large number of places where the derived class
"using" statement was needed in MLIR dialects, in addition to adding such
statements where needed in the ClangIR code.

The MLIR changes will likely be made upstream, but I am adding them here in the
clangir incubator as an proof-of-concept to show what will be needed in
downstream projects.