[CIR] Add option to emit MLIR in LLVM dialect. (#1316)
This PR adds the flag `-emit-mlir-llvm` to allow emitting of MLIR in the
LLVM dialect (cc @xlauko who asked me to do this).
I'm not sure if the naming of the flag is the best and maybe someone
will have a better idea.
Another solution would be to make the `-emit-mlir` flag have a value,
that specifies the target dialect (CIR/MLIR std dialects/LLVM Dialect).