[CIR][CUDA] CallConvLowering for basic types in NVPTX (#1468)
There are some subtleties here.

This is the code in OG:
```cpp
// note: this is different from default ABI
if (!RetTy->isScalarType())
  return ABIArgInfo::getDirect();
```
which says we should return structs directly. It's correct, has have the
same behaviour as `nvcc`, and it obeys the PTX ABI as well.
The comment dates back to 2013 (see [this
commit](https://github.com/llvm/llvm-project/commit/f9329ff650fa252606c65977405189dc46e5d66f)
-- it didn't provide any explanation either), so I believe it's
outdated. I didn't include this comment in the PR.