[CIR][NFC] Sort emit functions in CIRGenFunction.h
This change moves all declarations of emit* functions in CIRGenFunction.h
into a common location and sorts them alphabetically. The goal of this change
is to make it easier to keep upstream and incubator code in a consistent
location, making functions easier to find for upstreaming and minimizing
conflicts in the incubator when rebasing.

I did most of this sort manually, and I've probably been inconsistent in
how I treat sorting of uppercase versus lowercase. I made no attempt to
provide a rule for ordering different declarations of functions with the
same name. We can improve on that over time if anyone feels the need.

This change loses the grouping of some declarations that were co-located by
common purpose, but most of the declarations lacked a coherent ordering, so
I think this is a step forward overall.