[CIR] Fix for missing side effects during null pointer initialization (#1547)
When a pointer variable is initialized with a null pointer, the AST
contains a NullToPointer cast. If we just emit a null pointer of the
correct type, we will miss any side effects that occur in the
initializer. This change adds code to emit the initializer expression if
it is not a simple constant.

This results in an extra null pointer constant being emitted when the
expression has side effects, but the constant emitted while visiting the
expression does not have the correct type, so the alternative would be
to capture the emitted constant and bitcast it to the correct type. An
extra constant seems less intrusive than an unnecessary bitcast.