[CIR][NFC] Simplify struct type name creation
During review of a patch for upstreaming the cir.struct type support, Erich
Keane observed that the code we use for creating our type name for
structures with templates was likely to be error prone. He recommended
using getNameForDiagnostic instead.

After some experimentation, I found that with a couple of changes to the
printing policy, I could get the same strings we currently produce by
calling getNameForDiagnostic. This change does that.

Erich also pointed out that RecordDecls always have a DeclContext so
a few other lines could be eliminated where that was checked.