[CIR][ThroughMLIR] Lower structs and GetMemberOp
Structs are implemented as `memref<size x i8>`. It is not feasible to
represent them as tuples, for tuples can't be put in memref (i.e.
pointers to structs would break if we did).

We use `memref::ViewOp` for this. Unlike `PtrStrideOp`, the reinterpret
cast operation doesn't work here, as the result type is potentially
different from i8.