[CIR][NFC] Simplify struct type name creation (#1556)
During review of a patch for upstreaming the cir.struct type support,
Erich Keane observed that the code we use for creating our type name for
structures with templates was likely to be error prone. He recommended
using QualType::print with the appropriate printing policy instead. This
change does that.

Erich also pointed out that RecordDecls always have a DeclContext so a
few other lines could be eliminated where that was checked.