[CIR][NFC] Reverse the polarity of RecordType complete/incomplete
This change reverse the polarity of the cir::RecordType incomplete attribute,
making the attribute name "complete" and reversing all checks in the code
accordingly. The asm layout still prints "incomplete" for records that
are not complete and treats the absence of this text as meaning that
the definition provided is complete.

No functional change is intended.