[CIR] Treat cir.record class and struct types equivalently.
When we create a `cir.record` type, we track the `RecordKind` according to
whether the source code declared the type using `struct` or `class` even
though these are semantically equivalent. The distinction is used in naming
the type when we lower to the LLVM dialect.

This change updates the code to remove the last place where we were handling
class and struct records differently.