[CIR][CodeGen] Updates GlobalViewAttr's indices computation for the union type (#1584)
This PR introduces a new attribute for the explicit byte offset for
`GlobalViewAttr`. It's a proposal, so feel free to reject it once it's
not good from your point of view.

The problem is (as usually) with globals and unions: looks like we can
not really use indexes in the `GlobalView` to address an array of
unions. For example, the next program prints `4` now, but it should be
`42`:
```
typedef struct {
    long s0;
    int  s1;
} S;

typedef union {
   int  f0;
   S f1;
} U;

static U g1[3] = {{42},{42},{42}};
int* g2 = &g1[1].f1.s1;

int main() {       
   (*g2) = 4;
   printf("%d\n", g1[1].f0);    
   return0;
}
``` 
The problem is that we compute wrong indices in
`CIRGenBuilder::computeGlobalViewIndicesFromFlatOffset`. Maybe it can be
even fixed in this case, but I have a feeling that the fix would be a
bit fragile.

So instead of trying to support indexes for the array of unions I
suggest to use the offset explicitly.

From the implementation point of view there are some changes in
`CIRGenBuilder ` - but nothing really new is in there - I just did not
want to introduce copy-pasta for the `isOffsetInUnion` function that is
pretty the same as former `computeGlobalViewIndicesFromFlatOffset`.