[CIR] Remove ABI handling from CIRGen call handling (#1604)
Remove all the code the manages ABI info for arguments and return values
in the initial CIR codegen. We prefer for the CIR to represent the types
as they appear in the source code, with ABI handling being deferred
until the lowering phase or calling convention transform.

The ABI handling being removed here was brought over from the classic
codegen, but none of the effects being computed made it into the CIR so
this change is effectively NFC.

This change leaves the `CIRGenFunctionInfoArgInfo` with just one member.
That can be cleaned up in a later patch.