[CIR] Add code to detect non-zero-initializable records (#1603)
There are some cases where we're currently using the zero attribute to
initialize global variables for records that aren't properly
zero-initializable. We weren't checking for that, and we also weren't
properly calculating zero-initializability for records. This patch
addresses both of these issues.

This doesn't actually handle the case where the record isn't
zero-initializable. It just reports it as NYI. It does add a comment
about what needs to be done.