[CIR] Simlipify string literal global creation
Previously, when emitting a global for a string literal, we were creating
a GlobalOp, building a GlobalView attr for it, and looking up the global
from the symbol associated with the attr. This change splits out the
function that creates the global so that the global is returned directly
and the GlobalView attribute is only created in the case where it is needed.

This also updates the mechanism used for uniquing the global name used for
the strings so that if different base names are used the uniquing numbers
each base name separately. The mangling of the global used for strings is
not implemented, but the uniquing was happening prior to the mangling.
This change drops the uniquing below the placeholder for mangling.