[CIR] Generate SelectOp instead of TernaryOp for if cheap enough to evaluate unconditionally (#1642)
This came up during the review of
https://github.com/llvm/llvm-project/pull/138156

During codegen we check whether the LHS and RHS of the conditional
operator are cheap enough to evaluate uncondionally. Unlike classic
codegen we still emit `TernaryOp` instead of `SelectOp` and defer that
optimization to cir-simplify.

This patch changes codegen to directly emit `SelectOp` for `cond ?
constant : constant` expressions.