[CIR] Improved `cir::CastOp` verifier to allow bitcasts between types of the same size.
The `cir::CastOp::verify` method was overly conservative, and would fail on any `bitcast` from vector to scalar or scalar to vector.
This diff extends the `cir::CastOp::verify` method to check if the source and result types are the same size using the `mlir::DataLayout` of the current scope, and succeeds if the sizes match.
This diff also extends the CodeGen vectype tests with vector to scalar, scalar to vector and vector to vector conversions.
This diff also extends the IR invalid tests with vector to scalar and scalar to vector	conversions with different source and result sizes.