[CIR][CIRGen][Builtin][X86] Lower `vec_ext` related intrinsics (#1717)
Big question mark here:
When lowering target specific vector types: (`__m256i`, `__m128i`,
`__m64`), I was hitting an unreachable statement which I removed and
were preventing these types from being lowered. Not too familiar with it
but it's related to the attribute `"min-legal-vector-width"="N"` which
is not implemented for `cir::VectorType` as compared to OG. Is that a
blocker for these intrinsics as of now? or is that something we wanna
target before we merge x86 vector specific intrinsics?.