[CIR][CIRGen][Builtin][X86] Lower mm_prefetch (#1675)
Couple of things I have questions about:

1. I duplicated function `getIntValueFromConstOp` from
`CIRGenBuiltinAArch64.cpp`. I was wondering if that's correct or if
there's a place where we can avoid that duplication.

2. For the tests related to `mm_prefetch` im not sure if it'd be correct
to define them in a file eg: `sse-builtins.c` like it's currently done
in the codegen lib.

3. I'm also aware we can emit a call for a `PreFetchOp` would that be
required in this case?

related: https://github.com/llvm/clangir/issues/1414,
https://github.com/llvm/clangir/issues/1404 (A PR was previously opened
but It was not resolved)