[CIR][Lowering] Enabled VAArg lowering for `cir::PointerType` (#1762)
`cir::PointerType` was not included in the applicability guard for
`cir::VAArg` lowering during `LoweringPrepare`.
Since we don't have generic LLVM `cir::VAArgOp` (see [more
info](https://github.com/llvm/clangir/pull/1088#issuecomment-2465769265))
this causes an NYI error during lowering that doesn't need to happen.
To fix this I added the missing `cir::PointerType` to the `isa`. 
There is probably a more comprehensive fix to this if someone is
interested, this check should be removed and let the (possible) error
occur at the actual NYI site.